+++ src/components/icon/icon.test.jsx
@@ -4,47 +4,96 @@ import { render, screen } from '@testing-library/react';
 import Icon from './icon.jsx';
 
 describe('Icon', () => {
+  const requiredProps = {
+    name: 'foobar',
+  };
+
   it('has defaults', () => {
     const tree = renderer.create((
-      <Icon name="foobar" />
+      <Icon {...requiredProps} />
     )).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
-  it('has title', () => {
-    render((
-      <Icon name="foobar" title="yo" />
-    ));
-    expect(screen.getByTitle('yo')).toBeDefined();
+  describe('ariaLabel API', () => {
+    it('can be set', () => {
+      render((
+        <Icon {...requiredProps} ariaLabel="clear" />
+      ));
+      expect(screen.getByLabelText('clear')).toBeDefined();
+    });
+  });
+
+  describe('ariaLabelledBy API', () => {
+    it('can be set', () => {
+      render((
+        <Icon {...requiredProps} ariaLabelledBy="unique-id" />
+      ));
+      expect(screen.getByRole('img')).toHaveAttribute('aria-labelledby', 'unique-id');
+    });
+  });
+
+  describe('id API', () => {
+    it('can be set', () => {
+      render((
+        <Icon {...requiredProps} id="unique-id" />
+      ));
+      expect(screen.getByRole('img')).toHaveAttribute('id', 'unique-id');
+    });
+  });
+
+  describe('role API', () => {
+    it('can be button', () => {
+      render((
+        <Icon {...requiredProps} role="button" />
+      ));
+      expect(screen.getByRole('button'));
+    });
+
+    it('can be img', () => {
+      render((
+        <Icon {...requiredProps} role="img" />
+      ));
+      expect(screen.getByRole('img'));
+    });
   });
 
   describe('size API', () => {
     it('can be "small"', () => {
       render((
-        <Icon name="foobar" size="small" />
+        <Icon {...requiredProps} size="small" />
       ));
       expect(screen.getByRole('img')).toHaveClass('size-small');
     });
 
     it('can be "medium"', () => {
       render((
-        <Icon name="foobar" size="medium" />
+        <Icon {...requiredProps} size="medium" />
       ));
       expect(screen.getByRole('img')).toHaveClass('size-medium');
     });
 
     it('can be "large"', () => {
       render((
-        <Icon name="foobar" size="large" />
+        <Icon {...requiredProps} size="large" />
       ));
       expect(screen.getByRole('img')).toHaveClass('size-large');
+    });
+  });
+
+  describe('title API', () => {
+    it('can be set', () => {
+      render((
+        <Icon {...requiredProps} title="yo" />
+      ));
+      expect(screen.getByTitle('yo')).toBeDefined();
     });
   });
 
   describe('className API', () => {
     it('has a class name', () => {
       render((
-        <Icon name="foobar" size="large" className="i-am-a-classname" />
+        <Icon {...requiredProps} size="large" className="i-am-a-classname" />
       ));
       expect(screen.getByRole('img')).toHaveClass('i-am-a-classname');
     });
@@ -52,42 +101,42 @@ describe('Icon', () => {
     describe('fill', () => {
       it('can be "primary"', () => {
         render((
-          <Icon name="foobar" fill="primary" />
+          <Icon {...requiredProps} fill="primary" />
         ));
         expect(screen.getByRole('img')).toHaveClass('fill-primary');
       });
 
       it('can be "action"', () => {
         render((
-          <Icon name="foobar" fill="action" />
+          <Icon {...requiredProps} fill="action" />
         ));
         expect(screen.getByRole('img')).toHaveClass('fill-action');
       });
 
       it('can be "highlight"', () => {
         render((
-          <Icon name="foobar" fill="highlight" />
+          <Icon {...requiredProps} fill="highlight" />
         ));
         expect(screen.getByRole('img')).toHaveClass('fill-highlight');
       });
 
       it('can be "caution"', () => {
         render((
-          <Icon name="foobar" fill="caution" />
+          <Icon {...requiredProps} fill="caution" />
         ));
         expect(screen.getByRole('img')).toHaveClass('fill-caution');
       });
 
       it('can be "none"', () => {
         render((
-          <Icon name="foobar" fill="none" />
+          <Icon {...requiredProps} fill="none" />
         ));
         expect(screen.getByRole('img')).toHaveClass('fill-none');
       });
 
       it('can be "skip"', () => {
         render((
-          <Icon name="foobar" fill="skip" />
+          <Icon {...requiredProps} fill="skip" />
         ));
 
         expect(screen.getByRole('img')).not.toHaveClass('fill-primary');
@@ -102,42 +151,42 @@ describe('Icon', () => {
     describe('stroke', () => {
       it('can be "primary"', () => {
         render((
-          <Icon name="foobar" stroke="primary" />
+          <Icon {...requiredProps} stroke="primary" />
         ));
         expect(screen.getByRole('img')).toHaveClass('stroke-primary');
       });
 
       it('can be "action"', () => {
         render((
-          <Icon name="foobar" stroke="action" />
+          <Icon {...requiredProps} stroke="action" />
         ));
         expect(screen.getByRole('img')).toHaveClass('stroke-action');
       });
 
       it('can be "highlight"', () => {
         render((
-          <Icon name="foobar" stroke="highlight" />
+          <Icon {...requiredProps} stroke="highlight" />
         ));
         expect(screen.getByRole('img')).toHaveClass('stroke-highlight');
       });
 
       it('can be "caution"', () => {
         render((
-          <Icon name="foobar" stroke="caution" />
+          <Icon {...requiredProps} stroke="caution" />
         ));
         expect(screen.getByRole('img')).toHaveClass('stroke-caution');
       });
 
       it('can be "none"', () => {
         render((
-          <Icon name="foobar" stroke="none" />
+          <Icon {...requiredProps} stroke="none" />
         ));
         expect(screen.getByRole('img')).toHaveClass('stroke-none');
       });
 
       it('can be "skip"', () => {
         render((
-          <Icon name="foobar" stroke="skip" />
+          <Icon {...requiredProps} stroke="skip" />
         ));
 
         expect(screen.getByRole('img')).not.toHaveClass('stroke-primary');
@@ -152,42 +201,42 @@ describe('Icon', () => {
     describe('currentColor', () => {
       it('can be "primary"', () => {
         render((
-          <Icon name="foobar" currentColor="primary" />
+          <Icon {...requiredProps} currentColor="primary" />
         ));
         expect(screen.getByRole('img')).toHaveClass('color-primary');
       });
 
       it('can be "action"', () => {
         render((
-          <Icon name="foobar" currentColor="action" />
+          <Icon {...requiredProps} currentColor="action" />
         ));
         expect(screen.getByRole('img')).toHaveClass('color-action');
       });
 
       it('can be "highlight"', () => {
         render((
-          <Icon name="foobar" currentColor="highlight" />
+          <Icon {...requiredProps} currentColor="highlight" />
         ));
         expect(screen.getByRole('img')).toHaveClass('color-highlight');
       });
 
       it('can be "caution"', () => {
         render((
-          <Icon name="foobar" currentColor="caution" />
+          <Icon {...requiredProps} currentColor="caution" />
         ));
         expect(screen.getByRole('img')).toHaveClass('color-caution');
       });
 
       it('can be "transparent"', () => {
         render((
-          <Icon name="foobar" currentColor="transparent" />
+          <Icon {...requiredProps} currentColor="transparent" />
         ));
         expect(screen.getByRole('img')).toHaveClass('color-transparent');
       });
 
       it('can be "skip"', () => {
         render((
-          <Icon name="foobar" currentColor="skip" />
+          <Icon {...requiredProps} currentColor="skip" />
         ));
 
         expect(screen.getByRole('img')).not.toHaveClass('color-primary');