+++ CHANGELOG.md
@@ -5,8 +5,13 @@
     Click to see more.
   </summary>
 
+  - Major: remove unnecessary styling (inline-block and vertical alignment) on Icon component
   - Minor: Create the page patterns page.
   - Minor: Implement PageHeader component
+  - Minor: Implement Tag and TagSkill components
+  - Minor: Add ariaLabel to Icon props
+  - Minor: Add ariaLabelledBy to Icon props
+  - Minor: Add id to Icon props
 </details>
 
 ## 0.22.0 (May 22, 2020)